requirements.txt
boto==2.48.0
dj-database-url==0.4.1
Django==1.9.6
django-dotenv==1.4.1
django-nose==1.4.4
djangorestframework==3.6.3
gunicorn==19.5.0
nose==1.3.7
olefile==0.44
pep8==1.7.0
Pillow==4.2.1
psycopg2==2.7.1
pycodestyle==2.3.1
whitenoise==3.3.0
